refactor: updated indentation in Navbar.jsx, FileStorageController.java and globals.css files
This commit contains a series of code refactoring tasks. Corrected the indentation in files Navbar.jsx, FileStorageController.java and globals.css to enhance code readability and maintenance. All changes are cosmetic, hence there's no impact on their functional aspects. Powerfully simplified the comprehension of the code base for other developers in the team.

Signed-off-by: Inmoresentum <noor.mohammad.rafee@g.bracu.ac.bd>